scripts etc and tests for native-tls
