+++ client/src/tls/tests.rs
@@ -16,23 +16,14 @@ use std::sync::Arc;
 use std::sync::atomic;
 
 use futures::Stream;
-use native_tls;
-#[cfg(target_os = "linux")]
-use native_tls::backend::openssl::*;
-use native_tls::TlsAcceptor;
-#[cfg(target_os = "linux")]
-use openssl;
+use openssl::pkcs12::*;
 use openssl::pkey::*;
-#[cfg(target_os = "linux")]
-use openssl::ssl::{SSL_VERIFY_PEER, SSL_VERIFY_NONE, SSL_VERIFY_FAIL_IF_NO_PEER_CERT};
+use openssl::ssl::*;
 use openssl::x509::*;
-#[cfg(target_os = "linux")]
 use openssl::x509::store::X509StoreBuilder;
-#[cfg(target_os = "macos")]
-use security_framework::certificate::SecCertificate;
 use tokio_core::reactor::Core;
 
-use tls_native::{TlsStream, TlsStreamBuilder};
+use tls::{TlsStream, TlsStreamBuilder};
 
 use tests::tls::{root_ca, cert};
 
@@ -102,66 +93,84 @@ fn tls_client_stream_test(server_addr: IpAddr, mtls: bool) {
     let send_recv_times = 4;
 
     // an in and out server
-    #[cfg(target_os = "linux")]
     let root_cert_der_copy = root_cert_der.clone();
 
-    let server_handle = thread::Builder::new().name("test_tls_client_stream:server".to_string()).spawn(move || {
-    let pkcs12 = native_tls::Pkcs12::from_der(&server_pkcs12_der, "mypass").expect("Pkcs12::from_der");
-    let mut tls = TlsAcceptor::builder(pkcs12).expect("build with pkcs12 failed");
+    let server_handle = thread::Builder::new()
+        .name("test_tls_client_stream:server".to_string())
+        .spawn(move || {
+            let pkcs12 = Pkcs12::from_der(&server_pkcs12_der)
+                .and_then(|p| p.parse("mypass"))
+                .expect("Pkcs12::from_der");
+            let mut tls = SslAcceptorBuilder::mozilla_modern(SslMethod::tls(),
+                                                             &pkcs12.pkey,
+                                                             &pkcs12.cert,
+                                                             &pkcs12.chain)
+                    .expect("mozilla_modern failed");
 
-    #[cfg(target_os = "linux")]
-    {
-      let mut openssl_builder = tls.builder_mut();
-      let mut openssl_ctx_builder = openssl_builder.builder_mut();
+            {
+                let mut openssl_ctx_builder = tls.builder_mut();
 
-      let mut mode = openssl::ssl::SslVerifyMode::empty();
+                let mut mode = SslVerifyMode::empty();
 
-      // FIXME: mtls tests hang on Linux...
-      if mtls {
-          mode = SSL_VERIFY_PEER | SSL_VERIFY_FAIL_IF_NO_PEER_CERT;
-        
-        let mut store = X509StoreBuilder::new().unwrap();
-        let root_ca = X509::from_der(&root_cert_der_copy).unwrap();
-        store.add_cert(root_ca).unwrap();
-        openssl_ctx_builder.set_verify_cert_store(store.build()).unwrap();
-      } else {
-        mode.insert(SSL_VERIFY_NONE);
-      }
+                // FIXME: mtls tests hang on Linux...
+                if mtls {
+                    mode = SSL_VERIFY_PEER | SSL_VERIFY_FAIL_IF_NO_PEER_CERT;
 
-      openssl_ctx_builder.set_verify(mode);
-    }
+                    let mut store = X509StoreBuilder::new().unwrap();
+                    let root_ca = X509::from_der(&root_cert_der_copy).unwrap();
+                    store.add_cert(root_ca).unwrap();
+                    openssl_ctx_builder
+                        .set_verify_cert_store(store.build())
+                        .unwrap();
+                } else {
+                    mode.insert(SSL_VERIFY_NONE);
+                }
 
-    // FIXME: add CA on macOS
-    
-    let tls = tls.build().expect("tls build failed");
+                openssl_ctx_builder.set_verify(mode);
+            }
 
-    // server_barrier.wait();
-    let (socket, _) = server.accept().expect("tcp accept failed");
-    socket.set_read_timeout(Some(std::time::Duration::from_secs(5))).unwrap(); // should recieve something within 5 seconds...
-    socket.set_write_timeout(Some(std::time::Duration::from_secs(5))).unwrap(); // should recieve something within 5 seconds...
+            // FIXME: add CA on macOS
 
-    let mut socket = tls.accept(socket).expect("tls accept failed");
+            let tls = tls.build();
 
-    for _ in 0..send_recv_times {
-      // wait for some bytes...
-      let mut len_bytes = [0_u8; 2];
-      socket.read_exact(&mut len_bytes).expect("SERVER: receive failed");
-      let length = (len_bytes[0] as u16) << 8 & 0xFF00 | len_bytes[1] as u16 & 0x00FF;
-      assert_eq!(length as usize, TEST_BYTES_LEN);
+            // server_barrier.wait();
+            let (socket, _) = server.accept().expect("tcp accept failed");
+            socket
+                .set_read_timeout(Some(std::time::Duration::from_secs(5)))
+                .unwrap(); // should recieve something within 5 seconds...
+            socket
+                .set_write_timeout(Some(std::time::Duration::from_secs(5)))
+                .unwrap(); // should recieve something within 5 seconds...
 
-      let mut buffer = [0_u8; TEST_BYTES_LEN];
-      socket.read_exact(&mut buffer).unwrap();
+            let mut socket = tls.accept(socket).expect("tls accept failed");
 
-      // println!("read bytes iter: {}", i);
-      assert_eq!(&buffer, TEST_BYTES);
+            for _ in 0..send_recv_times {
+                // wait for some bytes...
+                let mut len_bytes = [0_u8; 2];
+                socket
+                    .read_exact(&mut len_bytes)
+                    .expect("SERVER: receive failed");
+                let length = (len_bytes[0] as u16) << 8 & 0xFF00 | len_bytes[1] as u16 & 0x00FF;
+                assert_eq!(length as usize, TEST_BYTES_LEN);
 
-      // bounce them right back...
-      socket.write_all(&len_bytes).expect("SERVER: send length failed");
-      socket.write_all(&buffer).expect("SERVER: send buffer failed");
-      // println!("wrote bytes iter: {}", i);
-      std::thread::yield_now();
-    }
-  }).unwrap();
+                let mut buffer = [0_u8; TEST_BYTES_LEN];
+                socket.read_exact(&mut buffer).unwrap();
+
+                // println!("read bytes iter: {}", i);
+                assert_eq!(&buffer, TEST_BYTES);
+
+                // bounce them right back...
+                socket
+                    .write_all(&len_bytes)
+                    .expect("SERVER: send length failed");
+                socket
+                    .write_all(&buffer)
+                    .expect("SERVER: send buffer failed");
+                // println!("wrote bytes iter: {}", i);
+                std::thread::yield_now();
+            }
+        })
+        .unwrap();
 
     // let the server go first
     std::thread::yield_now();
@@ -173,10 +182,6 @@ fn tls_client_stream_test(server_addr: IpAddr, mtls: bool) {
     // TODO: add timeout here, so that test never hangs...
     // let timeout = Timeout::new(Duration::from_secs(5), &io_loop.handle());
 
-    #[cfg(target_os = "macos")]
-    let trust_chain = SecCertificate::from_der(&root_cert_der).unwrap();
-
-    #[cfg(target_os = "linux")]
     let trust_chain = X509::from_der(&root_cert_der).unwrap();
 
 
@@ -191,13 +196,20 @@ fn tls_client_stream_test(server_addr: IpAddr, mtls: bool) {
     let (stream, sender) = builder.build(server_addr, subject_name.to_string(), io_loop.handle());
 
     // TODO: there is a race failure here... a race with the server thread most likely...
-    let mut stream = io_loop.run(stream).ok().expect("run failed to get stream");
+    let mut stream = io_loop
+        .run(stream)
+        .ok()
+        .expect("run failed to get stream");
 
     for _ in 0..send_recv_times {
         // test once
-        sender.send((TEST_BYTES.to_vec(), server_addr)).expect("send failed");
-        let (buffer, stream_tmp) =
-            io_loop.run(stream.into_future()).ok().expect("future iteration run failed");
+        sender
+            .send((TEST_BYTES.to_vec(), server_addr))
+            .expect("send failed");
+        let (buffer, stream_tmp) = io_loop
+            .run(stream.into_future())
+            .ok()
+            .expect("future iteration run failed");
         stream = stream_tmp;
         let (buffer, _) = buffer.expect("no buffer received");
         assert_eq!(&buffer, TEST_BYTES);
@@ -213,13 +225,19 @@ fn config_mtls(root_pkey: &PKey,
                root_name: &X509Name,
                root_cert: &X509,
                builder: &mut TlsStreamBuilder) {
-    // signed by the same root cert
-    let client_name = "resolv.example.com";
-    let (_ /*client_pkey*/, _ /*client_cert*/, client_identity) =
-        cert(client_name, root_pkey, root_name, root_cert);
-    let client_identity =
-        native_tls::Pkcs12::from_der(&client_identity.to_der().unwrap(), "mypass").unwrap();
-
     #[cfg(feature = "mtls")]
-    builder.identity(client_identity);
+    {
+        // signed by the same root cert
+        let client_name = "resolv.example.com";
+        let (_ /*client_pkey*/, _ /*client_cert*/, client_identity) =
+            cert(client_name, root_pkey, root_name, root_cert);
+
+        let client_identity = Pkcs12::from_der(&client_identity)
+            .and_then(|p| p.parse("mypass"))
+            .expect("Pkcs12::from_der");
+        let client_identity = Pkcs12::from_der(&client_identity.to_der().unwrap(), "mypass")
+            .unwrap();
+
+        builder.identity(client_identity);
+    }
 }
