+++ tests/ca.pem
@@ -1,0 +1,31 @@
+-----BEGIN CERTIFICATE-----
+MIIFTjCCAzYCCQCvf4GmlxfptDANBgkqhkiG9w0BAQsFADBpMQswCQYDVQQGEwJV
+UzETMBEGA1UECAwKQ2FsaWZvcm5pYTEWMBQGA1UEBwwNU2FuIEZyYW5jaXNjbzES
+MBAGA1UECgwJVFJ1c3QtRE5TMRkwFwYDVQQDDBByb290LmV4YW1wbGUuY29tMB4X
+DTE3MDQwNjA3MDUwMFoXDTE4MDQwNjA3MDUwMFowaTELMAkGA1UEBhMCVVMxEzAR
+BgNVBAgMCkNhbGlmb3JuaWExFjAUBgNVBAcMDVNhbiBGcmFuY2lzY28xEjAQBgNV
+BAoMCVRSdXN0LUROUzEZMBcGA1UEAwwQcm9vdC5leGFtcGxlLmNvbTCCAiIwDQYJ
+KoZIhvcNAQEBBQADggIPADCCAgoCggIBALBOeDzkUDa3dV1KOZhN/4WVMRhabHHu
++0dTUvfIkKPzlADB1DGLueEYXoulc3kQsrxJb6VosKuW5VRzKHU1/fJ8yAfmKh1I
+3vo9y55jEWPzm2Dvt6sXWCF++xrx6rkWAwdafXf2vo66fWezmmigVWqpaCxTR5em
+XqhpZkuf4R5aB6d99E40JQiEMs3R8tHLfRWvhGV7XBGdFt0baqShzaBnSPxV4buH
+FGjR3ZnbxInoHY7Mc36xxQArRhciG8KaJwyExKKCt9h2fbwK+evPXm0jCOmR05jE
+/pZCcXS+cu7cxRgIj46kIHoGnY1mLmhyNe/GJ/B8ukryYi+EtPU3nDci/oTo77fj
+hrsFjLHCmQnGLt8p+MqcrWni7UJ03bd/RATyuUP/ztuI5u9gBCP45rqgRNbS6MaX
+v5gXjxMU+x7oIBsVlfbkvUvrzwo1mGf6z3Lt5ctgrUgGfUsHvg9ad9K2rOFdExdG
+/zseX5WFXfCh6RxtzGAm+kBMd25KV2KvVekn7T3A09Ay7xIqGkou5lpxyIWs2RhT
+E0yohNiGuUDxopBYtMv6c7KhNybFZWE0m58+S07HUsPDp2gA8yk/1/YkacyAewlL
+GExYc0W3Vv3EVz6AasVTkDJW1BbwywzfMSW32JyT1EfeHW2XFYERa8FQkyYKRf3D
+hpn/7/nEiZpNAgMBAAEwDQYJKoZIhvcNAQELBQADggIBAIcAwR2KEyRCe1D4mb9I
+svuBKU5H6h9Tu2OyVfp/WrPVv7ooAAdjC1czzniwCdg04LbmvySWzijPCi7dNFLp
+huJIyvPdjPVRPXU7MqJR2djEgp7ppV4lIr8i3uPDhLLEZXbQmLWTacoZC+pFbUNq
+r6NBIA9I6uC67ibhY2fK3IURrBuaM6SWtrxPlDDUs2UODWnW8laI1hA+S4+DSUAv
+By7e5yJWBkf4xDAqiXaVIkj0GRPDOluNwNntkG9nUw6Lbb4ueioVCEu6+sIwctTc
+uZxsmPbq7EJrGNB5euPpPKF14QDjh9uK8m0dxHqEF7pIek8Szgna97+p6xrOE2Ty
+Xhmz7afHZmQ9bsrwReMCc7obuQgYcJimQfgXpFvYNRHfJX6OUhfocBFmxJNMSkM3
+XY2Hpf8ej5MftIpnmG7/5l+LmDgedi/iXfG3Zlo1Myq3PAInASkhzXJufL6zAcxm
+e8JL0Ogp0duPq1jn33vPsTN990z3ZU0AfC6qDSfyRfy2CQqZqkckgNYNVuIr8QMT
+PBa35hihM3EiRhf9l6m8uGib6sqIlCPWYf+rmcVmQaUTjxzwl3MPYSKcY2hPoIV7
+tJQITLS/Gzeef+cVHdpKxQQ10q2ud/wJWp+qSgzZSjnkoZkLk00GHUZleLnhRVdb
+WBZ38W160BdgsHNX4f9lDRFz
+-----END CERTIFICATE-----
