+++ tests/cert.pem
@@ -1,0 +1,32 @@
+-----BEGIN CERTIFICATE-----
+MIIFiDCCA3CgAwIBAgIJAIdx973umC+mMA0GCSqGSIb3DQEBCwUAMGkxCzAJBgNV
+BAYTAlVTMRMwEQYDVQQIDApDYWxpZm9ybmlhMRYwFAYDVQQHDA1TYW4gRnJhbmNp
+c2NvMRIwEAYDVQQKDAlUUnVzdC1ETlMxGTAXBgNVBAMMEHJvb3QuZXhhbXBsZS5j
+b20wHhcNMTcwNDA2MDcwNTAwWhcNMTgwNDA2MDcwNTAwWjBnMQswCQYDVQQGEwJV
+UzETMBEGA1UECAwKQ2FsaWZvcm5pYTEWMBQGA1UEBwwNU2FuIEZyYW5jaXNjbzES
+MBAGA1UECgwJVFJ1c3QtRE5TMRcwFQYDVQQDDA5ucy5leGFtcGxlLmNvbTCCAiIw
+DQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIBALTNUgesiIaPmzEuiBP9Pjgm/bpn
+rthPftBCclTxGfjMgEC6S2pFqHQxt2ZnFVKx385cmuJ/wGOQOujMzPSj9GYzpeCd
+wlzsSTbCOtAklTWcDSPqPdMO2rFxa24nrCsgxLjuAO1pjfzIgWZ+6nKQpMgWVQk7
+3EXtzZfqwxSlRq4B8DrfKatMvQJ4mQlLK61uhhDqBaeG0g9Y5IAD7ylbv/uGBJ7h
+18tUwHea+PYpEE+eADWZjAxDMn1KMzG25DfvXrYbIRmm63ebOOnVvPLPemwqyAXx
+X5sEbF+YZ6dFVgmZWOpqpicvHuRjdXt+eZB/WQC3JoLaciedOvcPwgC2j01OE0PX
+0kRv+dDzPPo5QDQNt05hu+esKR01LA9Pg7PBeYLbW2Ia2wf7pEwcKPUr8I4996Ph
+eZh+RqOj2ZxGpGJTvI8JoAL2M5MiwGPqWpXHl/5iBL7o7XdemPZlVdSJxahnkqMm
+55SZxfS5LTLeRZGXw08gDC7Ihrw2LYHH2I7JgOF4gLdeu4LuxdaAGL1ah8S/CFgO
+8StVdEBpTs23IJk3+6xK1BmKeRn+N5fObitUBIWzb+bsydXwyZHcEd4NhxPVGq/Q
+BA269eG90PWmz16JdMLbuoTplnPYisDuaFqYz99zoM1vXDgNUEszYlUDTO2d4Z7L
+MWglkfTc2W+77/QpAgMBAAGjNTAzMAkGA1UdEwQCMAAwCwYDVR0PBAQDAgXgMBkG
+A1UdEQQSMBCCDm5zLmV4YW1wbGUuY29tMA0GCSqGSIb3DQEBCwUAA4ICAQBMTEnd
+Q1tsD7gsAzDkFdL8ALZ1Dt4acDgcm3lpaWYUPIMeGPHYpYwjBt9WsT9KX45vW16b
+qv7DG4i1cRPaJF9HZuycA3axw12MzeNcqNe1bs5NgtJ2mXC+WibCCz8CrfYrKzlx
+YOqCbco5PbnCmtcsosawtQo251X+kwuTueFVXLurvmD0FZsmCzTxiM9e0XgYZ5J5
+jFvWV2z3M12fq2Zw6RIrWKijQOumQDBTYaOzONtmUtesBPgT9LK44muy+aXQErkz
+7tdkqqUCXywh1trIt4M+WJGZNhZfSMqyF+KivY0jQr6jfAkwgl8ZoriCdzIqeLZY
+PntpD88Zf5PfSpPJulTGe+GkLphBHS/g3vfmwc5tJHYd/f2X4AMXf/1Ofiq2yci5
+ILJTKrCh1eJEae07UUP9HQFSzrIyiD801ujHInYdKhZxbsFCm49D5C+BPmKD/3iW
+4yb+kxNhuulI54kKQr/K0bSLyNqr9Ks9UZqUxLnvUiaeWUMwIVQUXrw4XQUlBX5q
+OtunQDRUTfQ+89kftrQb7comuSN/+NaY7KbCEsBrWqf9LhXUOwb5J9ZooYm0r2h6
+jzekPplgCdy2qCAo1YcFBqD3rAHdFr5330NRDGrdBqm6Ee/JKxtIlyd8A+v7QoRk
+prjSS8T1B6qwbHdC4rgBRf+4YLrnzBBk42w1AQ==
+-----END CERTIFICATE-----
